Logo animaion
Finally added the logo animation, took forever. Had problems with the lottie library interpreting the file. After experimenting I have come to the conclusion that a https://lottiefiles.com/preview export is the best way to aproach the problem. The problem lies within the lottie.js player and not the bodymovin plugin in aftereffects (even though you can't preview) so use the lottiefiles.com link